fix: fixed actions on relation managers & move attach image to table action
