+++ app/Filament/RelationManagers/Wiki/ImageRelationManager.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace App\Filament\RelationManagers\Wiki;
 
-use App\Filament\Actions\Models\Wiki\AttachImageAction;
 use App\Filament\RelationManagers\BaseRelationManager;
 use App\Filament\Resources\Wiki\Image as ImageResource;
+use App\Filament\TableActions\Models\Wiki\Image\AttachImageTableAction;
 use App\Models\Wiki\Image;
 use Filament\Forms\Form;
 use Filament\Tables\Table;
@@ -102,7 +102,7 @@ abstract class ImageRelationManager extends BaseRelationManager
             ...parent::getHeaderActions(),
             ...ImageResource::getTableActions(),
 
-            AttachImageAction::make('attachimage'),
+            AttachImageTableAction::make('attachimage'),
         ];
     }
