+++ app/Filament/Resources/Wiki/Studio.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Filament\Resources\Wiki;
 
-use App\Filament\Actions\Models\Wiki\AttachImageAction;
 use App\Filament\Actions\Models\Wiki\Studio\AttachStudioResourceAction;
 use App\Filament\Actions\Models\Wiki\Studio\BackfillStudioAction;
 use App\Filament\Components\Columns\TextColumn;
@@ -241,8 +240,6 @@ class Studio extends BaseResource
 
             ActionGroup::make([
                 BackfillStudioAction::make('backfill-studio'),
-
-                AttachImageAction::make('attach-studio-image'),
 
                 AttachStudioResourceAction::make('attach-studio-resource'),
             ])
